requirements.txt
# Core
streamlit>=1.32.0
python-dotenv>=1.0.1

# Azure OpenAI
openai>=1.12.0

# LangChain - Updated to compatible newer versions
langchain>=0.1.0
langgraph>=0.0.40
langsmith>=0.1.40
langchain-core>=0.1.0
langchain-community>=0.0.32

# Data handling
pyyaml>=6.0.1
pydantic>=2.6.4

# Utilities
typing-extensions>=4.10.0
